Support finding a user by their email address rather than name
It is possible that a static or root user for the plugin would be configured using the email address of the user rather than the username - as Redis Enterprise will happily authenticate based on the email rather than username. If the email address is used, then `FindUserByName` would fail to find the user as it only checked the name, so `UpdateUser` would fail, and Vault wouldn't be able to rotate root user credentials.